requirements.txt
asgi-redis==1.4.1
asgiref==1.1.2
attrs==17.4.0
autobahn==17.6.1
Automat==0.6.0
backports-abc==0.5
certifi==2017.4.17
channels==1.1.5
chardet==3.0.4
click==6.7
constantly==15.1.0
coverage==4.0.3
daphne==1.3.0
Django==1.11.2
django-allauth==0.32.0
django-webpack-loader==0.5.0
djangorestframework==3.6.3
Flask==0.12.2
Flask-Cors==3.0.3
funcsigs==1.0.2
hyperlink==17.2.0
idna==2.5
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
msgpack-python==0.4.8
numpy==1.14.0
oauthlib==2.0.2
opencv-python==3.4.0.12
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
pytest==3.4.1
pytest-asyncio==0.6.0
pytest-django==3.1.2
python-coveralls==2.9.1
python-magic==0.4.15
python-openid==2.2.5
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.18.1
requests-oauthlib==0.8.0
singledispatch==3.4.0.3
six==1.11.0
tornado==4.5.3
Twisted==17.5.0
txaio==2.8.0
urllib3==1.21.1
Werkzeug==0.14.1
zope.interface==4.4.2
